requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.8.1=py36_0
astor=0.8.0=py36_0
blas=1.0=mkl
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2019.11.27=0
certifi=2019.11.28=py36_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cupti=10.1.168=0
cycler=0.10.0=py36_0
dbus=1.13.12=h746ee38_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
gast=0.3.2=py_0
glib=2.63.1=h5a9c865_0
google-pasta=0.1.8=py_0
grpcio=1.14.1=py36h9ba97e2_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.9.0=py36h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=h9c2bf20_1
intel-openmp=2019.4=243
jpeg=9b=h024ee3a_2
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py36he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.11.2=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
markdown=3.1.1=py36_0
matplotlib=3.1.1=py36h5429711_0
mkl=2019.4=243
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36ha843d7b_0
mkl_random=1.1.0=py36hd6b4f25_0
ncurses=6.1=he6710b0_1
numpy=1.17.4=py36hc1035e2_0
numpy-base=1.17.4=py36hde5b4d6_0
openssl=1.0.2u=h7b6447c_0
pcre=8.43=he6710b0_0
pip=19.3.1=py36_0
protobuf=3.11.2=py36he6710b0_0
pyparsing=2.4.5=py_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.5=hc3d631a_2
python-dateutil=2.8.1=py_0
pytz=2019.3=py_0
qt=5.9.6=h8703b6f_2
readline=7.0=h7b6447c_5
scipy=1.3.2=py36h7c811a0_0
setuptools=44.0.0=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.13.0=py36_0
sqlite=3.30.1=h7b6447c_0
tensorboard=1.14.0=py36hf484d3e_0
tensorflow=1.14.0=gpu_py36h3fb9ad6_0
tensorflow-base=1.14.0=gpu_py36he45bfe2_0
tensorflow-estimator=1.14.0=py_0
tensorflow-gpu=1.14.0=h0d30ee6_0
termcolor=1.1.0=py36_1
tk=8.6.8=hbc83047_0
tornado=6.0.3=py36h7b6447c_0
werkzeug=0.16.0=py_0
wheel=0.33.6=py36_0
wrapt=1.11.2=py36h7b6447c_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
